JAVA-2186 Add the ability to parse/output a BasicDBList
